Refactor connection management to improve code organization
Move connection-related types and functions to a new `conn` package for clearer separation of concerns. Rename `ConnReq` to `Request` and update references to align with the new structure. Add `CloseNotifier` implementation in the `conn` package for better connection handling.
